Move telemetry module to weightslab/utils/telemetry.py
